backend/pyproject.toml
[project]
name = "insurance-api-adapter"
version = "0.1.0"
description = "FastAPI adapter for V1 to V2 insurance APIs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "httpx>=0.25.0",
    "python-json-logger>=2.0.7",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adapter"]